New file for TCPServersHolder
